Rename user and theme reducers file names and move them to reducers folder
