+++ src/redux/reducers/user.reducer.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { User } from '../types/User';
+import { User } from '../../types/User';
 
 const initialUserState: User = {
     firstName: '',
